terraform: deploy echo container into kind cluster
This commit adds a basic Terraform module that deploys a Kubernetes
service running `hashicorp/http-echo` into a Kubernetes cluster. The
intent is to support deploying into a `kind` cluster using local
Terraform state storage for dev and CI, and clusters hosted in GKE with
remote Terraform state storage in GCS for staging and prod environments.
We crib enough of the `Makefile` from `prio-server` to achieve this,
with some modifications to `chdir` into the appropriate directory to use
the `kind` or cloud specific `main.tf`.