janus_server: implement `[leader]/upload`
Adds an implementation of an upload endpoint to `src/aggregator.rs`
wired up through a `warp` filter. We check report nonces and HPKE config
IDs per the PPM specification. We also decrypt the leader input share
while handling upload, primarily to exercise HPKE message decryption. We
might choose to not try to do that until the prepare sub-protocol begins
because there's no way to alert the client that decryption failed.

To test error conditions around report nonces, I also introduced a
`time` module which implements a simple clock interface.

We also extend the `aggregator` binary target so that it can be
instructed to run either a helper or a leader by passing the role name
as a positional argument (i.e., `./aggregator leader` or `./aggregator
helper`).

This commit does not implement the client side of upload (except
inasmuch as was needed to unit test `mod aggregator`; that will arrive
in a later PR.

Part of #12