Initial Datastore implementation.
Datastore is backed by a Postgres database. Tests are implemented via
the testcontainers crate, which allows spinning up a real instance of
Postgres. All transactions are currently fully serializable; I think we
should try very hard to maintain this, relaxing this constraint only if
we encounter a performance problem that cannot be solved through other
means.