Updates to message wire format, authenticated encoding/decoding. (#26)
* Update wire-format of AggregateReq messages.

This matches the new format introduced in the recent "interop nits" PR
(https://github.com/abetterinternet/ppm-specification/pull/206), and
will permit a much better MAC-validation implementation (which will be
the next commit).

* Introduce AuthenticatedEncoder/Decoder for message-encoding.

This replaces the existing ParameterizedEncode/Deocde<hmac::Key>-based
implementations, and was made possible by the recent "interop nits" PR
that moved task ID to be the first element of all authenticated
messages. This is a logically-required change: the correct key to use to
perform decoding was based on the task ID included in the message, but
retrieving the task ID required parsing the message, which previously
required decoding the message.

The new interface is similar to, but can't be expressed as, a
ParameterizedEncode/Decode trait implementation. We can't use those
traits because we have to have the entire message to perform
encoding/decoding (since the tag's byte offset is based on the end of
hte message), and the existing traits are based on the idea that there
may be additional data in the buffer after the message we are
encoding/decoding.

The expected usage of the decoder in application-level code is:
1. Read the entire message, e.g. as an HTTP request/response body.
2. Wrap the message in an AuthenticatedDecoder.
3. Call AuthenticatedDecoder::task_id() to determine the task ID
   associated with the message.
4. Look up the appropriate auth key based on the task ID.
5. Call AuthenticatedDecoder::decode() with the key retrieved in the
   previous step to retrieve the application-level message.
6. Continue application-level processing from there with the message
   retrieve in the previous step.

* Defer to HPKE library in use for standardized constants.