janus_server: introduce `struct TaskParameters`
Define a struct for PPM task parameters, and use it to create
`client::Client` instead of a handful of different parameters. This also
updates `datastore` to use a `TaskParameter` instead of a `TaskId`,
though I got stuck figuring out how to render all the fields of
`TaskParameters` into a prepared statement.