janus_server: write TaskParameters to database
Adds support for writing all the fields of a `struct TaskParameters`
into the `tasks` table in the database, with new fields/columns for the
aggregator auth key, HPKE config and HPKE private key. `tasks.ord` is
also now `role` and is an enumerated type with cases for the various PPM
protocol roles. Also adds `client::ClientParameters` to represent the
client's view of a PPM task, without secret values the aggregators need.

Part of #48