janus_server: rename subscriber installer fns
`trace::install_subscriber` is now `install_trace_subscriber` to make it
more obvious what the subscriber subscribes to, and
`trace::test_util::install_trace_subscriber` is now
`install_test_trace_subscriber` to make it easier to distinguish between
it and the subscriber intended for non-test cases.

Also installs a test subscriber in a few tests that failed to do so.