Factor prep_init failure to its own test, add prep_step failure test.
As noted during code review, the existing test case for prep_init was
not actually testing the right thing.

I also added a test_util::fake::Vdaf which is a minimal "fake" VDAF that
doesn't actually aggregate anything. It allows error responses to be
specified by the user. Right now the interface is exactly what is needed
for the current tests; since this is a testing utility only visible to
the crate's tests, I think this is OK.