Hide primary keys from users of the datastore library.
As part of this, I refactor all tables to have an "artificial" primary
key, which is a BIGINT; internally the datastore code manages the
necessary joins. Code using the datastore uses the "natural" primary
keys.

This simplifies the client code, while keeping join indexes small (which
may help performance) and allowing flexibility in case the "natrual"
primary key for a given entity changes with future changes to the PPM
spec.