Make console-subscriber an optional dependency (#54)
`console-subscriber` and its dependents are responsible for a
significant amount of compilation time, and yet they are only usable
when compiling with a special RUSTFLAGS value. This puts support behind
a non-default Cargo feature. I measured a time savings of 50 seconds
(2:28 to 1:38) on a clean release build, so this should give us a
noticeable win on CI job runtime.