Move EnvFilter to be per-layer, use Vec<Layer> (#61)
The upgrade to tracing-subscriber 0.3.10 brings with it some new
features for better ergonomics. First, EnvFilter can now be used as a
per-layer Filter, rather than only as a Layer. By using it as a Filter
on our stdout layers, we can now avoid modifying a global filter when
using console-subscriber. Instead, we put EnvFilter on the stdout layer,
and Targets on the console-subscriber layer, and the two do not
interfere with each other. Additionally, Vec<L: Layer> now implements
Layer, so we can create a vector of one or two boxed layers, rather than
handling four different Option variables. This will also greatly
simplify adding additional subscribers. (such as tracing-opentelemetry)