Implement datastore::Crypter, to be used to encrypt values.
datastore::Crypter will encrypt/decrypt values to be stored in a
datastore. The ciphertexts are cryptographically bound to the specific
database cell (i.e. table, column, & row) they are to be stored in.

N.B. on use of LessSafeKey: the alternative that ring seems to want me
to use is a combination of OpeningKey/SealingKey, which are AEAD keys
associated with a "nonce sequence." While using a nonce sequence makes
sense for e.g. a wire protocol that can determine the nonce from a
packet number or similar value, for our use-case it would be challenging
to encrypt with a specific sequence of nonces & impossible to decrypt.
Instead, I opt to use randomly-generated nonces, including the nonce
values in the ciphertexts.

I think using random nonces is safe in this context: nonces are 12 bytes
in length, giving 2^(12*8) = 2^96 possible nonce values. The birthday
bound then tells us we would expect to see a nonce collision after
generating around sqrt(2^96) = 2^48 nonces. Given that we are planning
to use this to encrypt a number of different plaintexts that is bounded
by about 3 times the number of tasks, I suspect we won't be anywhere
near this limit for quite some time.

Part of #48.