Aggregator: allow multiple keys for a given task ID.
Currently, the with_authenticated_request function takes an Aggregator.
That may change once we support multiple tasks, but the important bit is
that the keys reference lives as long as the parameter.

I also change with_authenticated_request to not include the decoding key
in its response, since it is no longer needed.