Store aggregator auth keys in a separate table.
This is consistent with task_hpke_keys, and will allow additional
per-key metadata to be added later as necessary. I decided to keep the
in-memory vectors of keys in the same order as is stored in the
datastore, so I changed how the aggregator uses the list of keys -- when
decoding we try keys in *reverse* order by ord, and when encoding we
generate the auth tag based on the *last* key by ord.