Aggregator support for multiple HPKE configs/keys. (#81)
The largest part of this PR is a refactoring of HpkeSender &
HpkeReceiver to be free functions seal/open instead. I think this is a
good refactoring because it binds the application info more directly to
the place it is used (i.e. we know from where we are in the code what
application info to use; passing around HpkeSender/Receiver would allow
potentially binding to _incorrect_ application data).