Add integration test for SIGTERM behavior
This test checks that the aggregator server exits promptly when it
receives a SIGTERM signal. By default, Kubernetes requests that
containers stop with this signal. Currently, the aggregator binary does
not override the default SIGTERM behavior, so it trivially passes this
test, but we may override this to give in-flight requests a chance to
complete.