Move ephemeral_datastore() to new crate, via macro (#95)
This moves the declaration of items necessary for an ephemeral datastore
into a macro in a new crate, `test_util`. This macro is then invoked
inside `janus_server` only in its unit tests, as well as separately in
each integration test. As a result, `testcontainers` can be removed from
`janus_server`'s runtime dependencies.