Helper `/aggregate_share` endpoint
Implements the helper's `aggregate_share` endpoint. The assumption is
that the process of preparing input shares into output shares will
create rows in `batch_aggregations` and update the `checksum` and
`aggregate_share` columns as individual reports are prepared. Then, all
the `/aggregate_share` handler has to do is sum the aggregate shares it
finds.

Note that this does not include support for the protocol changes in [1],
nor does it include enforcement of a task's `max_batch_lifetime`.

[1]: https://github.com/ietf-wg-ppm/ppm-specification/pull/224

Part of #105