Safer time calculations
We had a number of risky casts between `chrono::Duration` and
`janus_server::message::Duration`, or between `i64` and `u64` in other
places. With this commit, we consistently use `message::{Duration, Time,
Interval}` everywhere. We also now have a safe constructor for
`Interval` that forbids the creation of an interval whose end cannot be
represented as a `Time`. We also have safe routines for doing math on
durations and times that return errors if overflows occur.