Workaround: connect to ephemeral db on 127.0.0.1 (#108)
The `testcontainers` crate makes use of `docker run` with the
`--publish-all` flag to expose our Postgres database for connections. In
certain situations, Docker's port allocation assigns different port
numbers to the same container via the host's IPv4 and IPv6 addresses.
The `get_host_port()` method does not provide which host IP address a
port corresponds to, nor specify which address's port will be returned.

Empirically, it appears that Docker will return port bindings for
0.0.0.0 before ::, and `testcontainers` will select the first port
binding in the list. Thus, as a workaround, we use 127.0.0.1 in the
connection string, instead of "localhost", to keep the database client
from trying both socket addresses, and possibly connecting to the wrong
container.