Parameterize BatchUnitAggregation by vdaf::Aggregator.
Previously, it was parameterized by various messages, which was
unnecessarily flexible & somewhat confusing. I modified some tests to
work with Poplar1, to allow testing with a non-unit aggregation
parameter.

While working on this I noticed that
get_batch_unit_aggregations_for_task_in_interval assumes a deterministic
encoding for aggregation params, since it does an index lookup based on
the encoded aggregation parameter. I think this is true for all
aggregation parameters today (prio3 uses the unit type which certainly
qualifies; poplar1 uses BTreeSet which has a deterministic iteration
order). We'll need to ensure this remains true in the future to avoid
the potential for weird bugs -- I can't think of a (performant) way to
do better.