Datastore: refactor single-entity getters to return Option<Foo>.
This eliminates the NotFound error, which was previously used for this
purpose. I introduce a new error "MutationTargetNotFound" for when a
mutation can't find the row it is supposed to be mutating.

As part of this, I introduce a new UnrecognizedAggregationJob PPM
problem type. Previously, this case would have translated into an
internal server error, which is probably not what we want.