Initial implementation of aggregation job creator. (#118)
The concurrency part was the most challenging piece: I would have liked
to have the task-discovery worker (i.e. AggregationJobCreator::run()) be
"tightly linked" to the job-creation workers (i.e.
AggregationJobCreator::run_for_task()) in the sense that a
task-discovery worker shutting down would be noticed & cleaned up
immediately (in a select loop). Sadly, the lifetime/ownership issues
involved made this very challenging. Instead, I settled for a strategy
where the task-discovery worker owns a oneshot sender for each
job-creation worker, and is responsible to notice when a task is no
longer relevant; actual shutdown of the job-creation worker is loosely
coupled and may happen later. This approach is hypothetically racy and
definitely slower to notice a lingering no-longer-existent task, but I
think those issues are marginal enough that this is an acceptable
strategy -- I think the worst outcome we'll see is, if we delete a task,
we'll see its job-creation worker produce errors until the next time the
task-creation worker updates its view of existing tasks.