Initial split-out of common crate (#146)
This moves some of the most primitive types into a common `janus` crate.
In the future, this will be reused from separate client and server
crates. (with the possible addition of a collector crate as well)
Additionally, the methods of `Role` were made public, `MockClock` was
moved to the `test_util` crate, and a constructor was added to `TaskId`.