Fix Dockerfile entrypoint, pass args and use exec
This change to the ENTRYPOINT line fixes two problems. First, by using
exec, it makes the server PID 1 again, so that it can receive signals.
Second, it propagates command line arguments passed to the container
onward to the server. It also adds a simple CI test that the container
images can be run successfully.