client: remove type param, add concrete upload fns
This removes the type parameter on Client that specified a vdaf::Client,
instead using our crate::task::Vdaf enum to specify a VDAF
instantiation. Separate `upload_boolean()` and `upload_integer_u128()`
methods are added, which take a measurement of a concrete type as an
input, and perform dispatching based on the enum to the correct
implementation. (Adding an `upload_byte_string()` method for Poplar1 is
left as future work, for once the Poplar1 VDAF is better specified and
implemented)