Leader collect job URI handler
Adds a warp filter for path `/collect_jobs/{collect_job_id}` to the
leader. Adds support for querying collect jobs from the datastore as
well as updating them with helper and leader aggregate shares. The
latter is currently only needed for tests, but will soon be used when
running collect jobs.

Part of #105