Collect job driver
Fleshes out the implementation of the Janus collect job driver.

Some existing logic used for the helper's `/aggregate_share` handler is
refactored into `mod aggregate_share` so it can be used in
`collect_job_driver`.

Additionally, the existing methods `update_collect_job_*` methods on
`datastore::Transaction` are collapsed into a single method that sets
helper aggregate share, leader aggregate share, report count and
checksum in a single operation. This simplifies the logic of the collect
job driver since it doesn't have to deal with the case where the
leader's aggregate share was computed but the helper's isn't known yet.
The downside is that if a collect job fails because the helper failed to
compute its aggregate share, then the leader will recompute its share
"from scratch" the next time the collect job is run. If helpers fail
often enough to make caching the leader aggregate share worthwhile, then
we probably have bigger problems than this performance issue.

Part of #105