JobDriver: remove all non-required parameters from callbacks.
The callbacks (job-acquirer, job-stepper) now only receive information
that is managed/computed by the JobDriver itself; data that is
effectively "constant" that the JobDriver was holding onto is now
expected to be owned by the callback itself.

This change may be controversial because it pushes more logic into the
application rather than keeping it in the "helper" JobDriver. I think
it's worthwhile not just from a cleanliness perspective but also because
it opens the door to later cleanups/optimizations--as a minor example,
we could now place the datastore into the {Aggregation,Collect}JobDriver
and no longer need to manage an Arc<Datastore>.