Aggregation job driver: give up on jobs that fail too many times.
Failure is implemented as moving the aggregation job to the FINISHED
state. We don't change the state of any of the included report
aggregations, on the theory that we may want to debug the issue &
keeping state around will help with that; this is the only way that an
aggregation job can reach the FINISHED state with report aggregations in
a non-terminal state.