Update libprio-rs dependency to version 0.8.0.
This version of libprio-rs implements draft-irtf-cfrg-vdaf-01, which is
the version we are targeting for interop tests.

This change is large but mostly-mechanical. A large number of line
changes are due to the new `L` const generic parameter to the VDAF
Aggregator trait.  This parameter is effectively the underlying PRG's
seed size.  Unfortunately, Rust currently doesn't allow const generic
parameters to be inferred; there is an open issue to resolve this, but
it doesn't appear to have received much attention thus far[1].

[1] https://github.com/rust-lang/rust/issues/80528