Monitor completion of some async tasks in tests
This adds a new trait that abstracts over `tokio::task::spawn()`, and
adds a test-only implementor that monitors when spawned async tasks
complete. Tests can create multiple such `TestRuntime`s with different
labels, and wait for some number of tasks to be completed before
continuing. This addresses one of the two reasons we needed sleeps in
tests for. Those tests are rewritten with this new functionality.

`TestRuntime` provides one additional feature: if a task spawned through
it panics, then this will be detected, and the test's main task will
panic when a related object drops.