Add logic/DB query/tests for future Poplar support (#267)
* Add logic/DB query/tests for future Poplar support

This adds datastore and aggregation job creator methods to support VDAFs
with non-unit type aggregation parameters. For now, these are test-only
methods. When full Poplar support is added in the future, this can be
wired into existing code that dispatches upon `VdafInstance`.

Aggregate job creation is likely the place where Janus's behavior will
differ the most between Prio and Poplar, because the aggregation job
creator must react to new collect jobs, rather than new client reports.
Additionally, any one report may be placed in more than one aggregate
job, so that preparation can be run with multiple aggregation
parameters.

* Fix an inconsistent test fixture

This aggregation job was created with a different TaskId than a report
that is linked to it through a report aggregation. This inconsistency
did not affect the the outcome of the test, but fixing it should better
model true operation, and may avoid future issues.

* Move Aggregator trait bounds to where clause

* New issue for improvements to LIMIT 5000 queries

* Add marker trait for VDAFs with an agg param

* Expand upon datastore method doccomments

* Review comments