Build a reportTooEarly problem document for reports from the future.
(Note, there is already a upload_report_in_the_future test; however, it
calls the aggregator function directly, and so checks against the
internal aggregator::Error that is returned rather than validating the
problem document as e.g. aggregate tests do. We should arguably modify
these tests to hit the HTTP endpoints for realism of testing, but I am
not going to make that change as part of this commit.)