rename crate `janus` -> `janus_core`
Crate `janus_client` depends on `janus`, so if we want to publish the
former to crates.io, then the latter has to go, too. The problem is that
there is already a crate called `janus` up there. This commit renames
our crate to `janus_core` to avoid the conflict.

Part of #222