Publish SQL schema to GCS bucket
On release, we use `gcloud storage` to upload the SQL schema file to
a GCS bucket. We version the artifacts by uploading them under a
"directory" named after the git tag. This allows us to upload multiple
SQL files in the future and doesn't require the consumer of the SQL
files to parse out or otherwise ignore the version component.

Part of #222