Pin Kind node image used in EphemeralCluster
By default, `kind` would install a node image for Kubernetes 1.24, which
is a couple minor versions past what we deploy to cloud environments and
which runs afoul of some Terraform bugs[1]. Pin the Kind node image we
use to a Kubernetes 1.22 image, to match the Kind we use in other CI
contexts. We also make sure to install the `kind` CLI ourselves in the
CI workflow, to guarantee that the node image we pin comes from the same
release.

[1]: https://github.com/hashicorp/terraform-provider-kubernetes/issues/1724