Integration test: submit exactly one batch worth of measurements.
This works around an issue in Daphne where some measurements will not be
included in the final aggregate when Daphne is in the leader position,
causing the final aggregated value to be incorrect. I work around this
instead of disabling the test in order to show that most of the
functionality still works.