Daphne integration test. (#346)
Also, refactor the existing Janus-only integration test to use the same
code as the new Daphne-Janus integration tests.

The hermetic Daphne instances are implemented by including a compiled
version of Daphne as an artifact, and running it inside miniflare, based
heavily on the manual instructions in the Daphne repository. Miniflare
is now a required dependency for running tests; instructions for running
tests & updating the version of Daphne used for tests is included in the
README.md. Ideally, we would start a container, but Daphne does not yet
build out a suitable container.