Update integration test to serve Janus-Janus test out of /dap/ prefix.
This exercises a current bug in Janus where it can't interoperate with
an aggregator that is not serving out of the "base" of its domain (i.e.
"https://example.com/" will work, "https://example.com/dap/" will not).
The test currently fails; the following commit will fix this bug & make
the test pass.

This was missed because all existing tests of sufficient complexity to
potentially exercise a bug in aggregator endpoint handling were testing
against endpoints set to the base of a domain.