janus_core: specify version of k8s-openapi. (#386)
We don't actually use this dependency (directly) anywhere. It's pulled
in via `kube`; we only specify it so that we can specify the `v1_20`
feature. We previously specified a wildcard version since we just want
to match what `kube` pulls in. Unfortunately, crates.io disallows the
use of wildcard versions, which blocks us publishing 0.1.5 to crates.io:

```
error: failed to publish to registry at https://crates.io

Caused by:
  the remote server responded with an error: wildcard (`*`) dependency constraints are not allowed on crates.io. See https://doc.rust-lang.org/cargo/faq.html#can-libraries-use--as-a-version-for-their-dependencies for more information
```

Now, we just specify what is getting pulled in anyway. The downside is
that we'll need to do manual work to keep these in sync if `kube` ever
changes its effectively-major version.