Specify versions for dev dependencies, too.
This is a guess; I get the same error as in
https://github.com/divviup/janus/pull/386, but the only remaining "*"
dependencies are dev-dependencies. Unhelpfully, `cargo publish
--dry-run` does not encounter this error.

IMO this is a wart in cargo/crates.io: "*" dependencies aren't allowed
since it means "this will work with every version ever". But specifying
"*" is the only way to say "I don't care, use whatever version is
already included due to other dependencies." There's apparently no way
to add a feature to an already-included dependency and also publish to
crates.io.