Task: require/allow collector_auth_token iff we are in Leader role.
This matches how things will be configured: helpers don't have knowledge
of the leader-collector auth token.

Also, update aggregator endpoint checks to return a 404 if the endpoint
doesn't apply to the task role BEFORE checking auth. I think this makes
more sense since I suppose the desired semantics are "the resource does
not exist if the task is for the wrong role" and it doesn't make sense
to authenticate against a task that doesn't exist.