Run container builds in parallel
In a recent PR of mine, the `janus_server_docker` step in `ci-build.yml`
took nearly 45 minutes[1]. It takes 18 minutes to build the
`janus_server` container from `Dockerfile`, and then the builds for the
other Janus components are quite swift because the build cache gets
mounted in. But it doesn't appear to speed up the build of
`janus_interop_client`, which takes another 15-16 minutes.

In this commit, we break out the builds of the `janus_interop_client`,
`janus_interop_aggregator` and `janus_interop_collector` containers into
their own job, so that GitHub can run those in parallel with building
the other containers and with running `janus_server_build`.

[1]: https://github.com/divviup/janus/runs/7887936297?check_suite_focus=true