Compress Docker images with zstd.
Note that the `zstd` crate compiles the C zstd library. I think this is
OK since it is only used during build & tests, i.e. this does not
introduce a new non-Rust dependency into our release binaries. I went
with zstd as it is generally considered to have an excellent tradeoff in
compression speed & compression ratio. Testing the `zstd` CLI tools with
default configuration against the current image that is built, the image
is reduced from 751 MiB to 191 MiB, ~25% of the original size.
Anecdotally, I don't notice a different in build script time -- which
makes sense, as the CLI tool compressed the image in a few seconds.

Also, factor some of the more complex common logic between the Janus &
Daphne image-handling code out to common libraries. Since some of this
logic lived in build scripts, I needed to create a new crate
`build_script_utils` for the build scripts to depend on.