Build interop containers with a new "small" build profile.
This build profile is intended to produce minimally-sized binaries; this
change shrinks the compressed image from 191MiB to 144MiB. With this
profile, the actual binary is only 11MiB, so I think I'm running out of
room to shrink the binary itself.

This change also makes our ordinary builds of interop binaries use the
release build profile again, too, which I think is what we probably
want.