Container-based interop testing (#408)
We currently build a "shim" Daphne container based on a compiled
version already checked into our repository. This is temporary: I plan
to build a (pinned) version of Daphne soon using a similar technique.

We also specify `-p` to `cargo` when building release binaries, which
is needed to avoid the workspace feature convalescence feature from
enabling undesirable features on binaries because those features are
enabled elsewhere in the repository.