CI documentation step: write to /tmp.
To avoid potentially causing problems with our cache, we were already
writing to a "bogus" subdirectory of the build artifacts directory. I
_think_ this would still work OK with the new cache action, but no need
to risk it -- intead, now documentation is written to a subdirectory of
/tmp.

Also, move the documentation check up to before the expensive build/test
steps, so that we fail earlier & faster for these lint-style checks.