Skip building the container images for lints.
Specifically, `cargo clippy` was running the build script & therefore
building the container image, even though the intent was just to check
for errors without comipling the code. This required writing a little
code to allow building of the container to be skipped -- attempting to
skip building the container in an actual `cargo test` run will lead to
test failures.