Retries in HTTP clients
Adds a new module `janus_core::retries` which uses crate `backoff` to
allow retrying HTTP requests based on whether an error is transient. We
retry if we encounter:

  - failures to connect (i.e., connection refused or reset);
  - timeouts;
  - HTTP errors in the 500 range, except 501 Not Implemented.

The error classification depends on methods on `reqwest::Error` and
walking the chain of `std::error::Error::source` errors to see if there
is an `std::io::Error` under the covers.

We retry on HTTP status 500, 502-599, but if we give up after
encountering enough such errors, the caller gets
`Err(Ok(reqwest::Response))` so that they can examine the response body
or headers.

Resolves #196