Use containers for interop API tests.
With this commit, the `interop_binaries` E2E tests now test the interop
test containers, rather than the binaries that are added to the
containers. This is important since the container build definitions
include important wiring -- before this change, the container builds
could be broken (eg Janus can't talk to datastore), but the E2E tests
would still pass.