Don't trace error in `datastore::delete_task`
When `janus_cli` provisions a task, it first deletes any task with the
provided `TaskId`, and ignores the error if it is "Not Found". But if
the `delete_task` method is annotated with
`#[tracing::instrument(err)]`, then you get a big, scary level=ERROR
line printed in RED to the console when `janus_cli` is run, even though
the error is benign. Remove the annotation to avoid operators
incorrectly thinking something has gone wrong.